Remove repository nesting
