+++ template/.github/workflows/daily-scheduled-ci.yml.jinja
