+++ template/.github/workflows/{% if upload_conda_package or upload_pip_package %}pre-release.yml{% endif %}.jinja
