+++ template/docs/api/{% if command_line_interface %}cli.md{% endif %}.jinja
