+++ template/docs/static/hooks.py.jinja
