+++ template/src/{{module_name}}/__init__.py.jinja
