+++ template/src/{{module_name}}/core.py
