+++ template/src/{{module_name}}/{% if command_line_interface %}cli.py{% endif %}.jinja
