+++ template/tests/{% if command_line_interface %}test_cli.py{% endif %}.jinja
