+++ template/{% if create_author_file %}AUTHORS.md{% endif %}.jinja
