+++ template/{% if create_jupyter_notebook_directory %}examples{% endif %}/1_intro_to_{{module_name}}.ipynb.jinja
