+++ template/{% if open_source_license == 'Apache-2.0' %}LICENSE{% endif %}.jinja
