+++ template/{% if open_source_license == 'BSD' %}LICENSE{% endif %}.jinja
