+++ template/{% if open_source_license == 'GPL-v3' %}LICENSE{% endif %}.jinja
