+++ template/{% if open_source_license == 'MIT' %}LICENSE{% endif %}.jinja
