+++ template/{% if upload_conda_package %}conda.recipe{% endif %}/meta.yaml.jinja
